# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "Jix Express API Demo",
    "description": "A simple Express API Vertical Slice demo"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/api/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "activeOnly",
            "in": "query",
            "schema": {
              "type": "bool"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contactArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/contactInput"
              }
            }
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/contactInput"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "test123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "isDisabled": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "name"
        ]
      },
      "contactArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/contact"
        }
      },
      "contactInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "isDisabled": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "name"
        ]
      }
    }
  }
}